requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic-settings
pydantic[email]
numpy
alembic
psycopg2-binary
greenlet
PyJWT
python-multipart
httpx
apscheduler
google-genai
google-auth
